requirements.txt
requests
psycopg2-binary
beautifulsoup4
pypdf2
selenium
pymupdf
